Implemented Hash Table Recomendations
I've implemented the recommendation algorithm using a hash table. I have also extracted the movie struct from "GraphRecommend" to be in it's own header file. This allowed me to reuse the movie struct for the hash table. I have also updated the CMakeList.txt to include the new files when using CMake. I've compiled it using VSCode to ensure this was done correctly. The main function was modified to be able to select between graph and hash recommendations before asking for recommendations.